+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- docker: Set default settings of api server for SoC devices
 - cli: Env overwrites api server config settings
 
 ## [1.14.4] - 2023-12-17