+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- webapp: No browser history for singe media navigation
 - webapp: search tags with blanks in Detail
 
 ## [1.15.0] - 2024-01-27